app/chatbot/requirements.txt
fastapi
uvicorn
transformers
torch
accelerate
bitsandbytes
boto3
pydantic
peft
python-dotenv


